Abstract out the TitleView's implementation
By abstracting out the TitleView's implementation I am now able to drop
in a custom WebView version that allows for some nifty experience in
displaying HTML text.

This includes a refactor to the core where you can override the default
FSTitleView and a sample WebView implementation in the demo project.